backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
supabase>=2.0.0
anthropic>=0.19.0
twilio>=8.0.0
python-dotenv>=1.0.0
httpx>=0.24.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
